KMM: Loader: insert parsed game; navigate to game

+++ KMM/composeApp/src/androidMain/kotlin/com/letstwinkle/freebee/database/EntityIdentifier.kt
@@ -1,3 +1,3 @@
 package com.letstwinkle.freebee.database
 
-actual typealias EntityIdentifier = Int
+actual typealias EntityIdentifier = Long
